Extract Notification Class
refactor: Notification Class move
Moved notification class to its own file to simplify the project.

todo:
 - Comments
